agent/requirements.txt
ws4py
cherrypy
msgpack-python
pillow
argparse
